Extract mutation recursion logic into a Mutator class
